30th Commit- Currently working on FuelCapacity send to fuel fragment, coverableDistance and nextFuelFill.

+++ app/src/main/java/com/johnyhawkdesigns/a62_car_bike_fuelaveragecalculator/fragments/AddEditVehicleFragment.java
@@ -196,7 +196,7 @@ public class AddEditVehicleFragment extends DialogFragment {
                 vehicle.setVehicleType(vehicleType);
                 vehicle.setVehicleMake(vehicleMake);
                 vehicle.setVehicleModel(vehicleModel);
-                vehicle.setVehicleFuelCapacity(Double.parseDouble(vehicleFuelCapacity));
+                vehicle.setVehicleFuelCapacity(Integer.parseInt(vehicleFuelCapacity));
 
 
                 Log.d(TAG, "onClick: vehicleType = " + vehicleType + ", vehicleMake = " + vehicleMake + ", vehicleModel =" + vehicleModel);
